Change GMaps API init to occur on browser start.
As opposed to extension installation. This is because otherwise the
extension will have undefined variables for things like its geocoder
object if the user installs the API, shuts down Chrome and then re-opens
Chrome sometime later. With this, we re-initialize such variables on
browser start so we are ready to go.