Rename building -> listing
Renamed variables and functions and etc that used the nomenclature of
'building' instead of 'listing' so that it is consistent and not as
confusing/misleading. Technically, these were listings, and not
necessarily buildings.

(Even though streeteasy seems to internally refer to them as whole,
entire buildings, which they often are not).