Refactor content script code into extendible OOP
Refactored the content script code to be extendable and structured. The
idea is now that if we wish to support another site (like
apartments.com) then we implement the ApartmentSite interface and
implement the necessary functions. Upon doing so, the rest of the code
follows the exact same logical sequence removing duplication. This
interface gives us structure by emphasizing on the similarities between
sites and letting the site-specific logic involving things like their
HTML isolated to each site and hidden from others.