Increase maps delay to 650 milliseconds
This is pretty unacceptable but its the best we can do for a problem
like this. The other alternative is to send multiple messages for each
address as if they were a single address and have each message keep
re-trying until it gets a successful result from GMaps, but this would
require some legwork in the backend. Issue #4 deals with this idea for a
solution.