Move destination global setting to a function
Previous, gmaps_util.js would execute some Chrome storage logic to get
values out of Chrome storage and set the destination globals with the
values. This was done whenever gmaps_util.js was loaded. However, this
was a fundamentally flawed strategy, since it relied on the code in
gmaps_util.js executing and finishing before any other dependent
Javascript code ran. In particular, this caused issues with apartment
sites if their apartment site-specific code ran before gmaps_util.js
could finish its initialization of the destination globals, which were
used in the apartment specific code. By moving the destination globals
initializations to a function, and then ensuring that the function is
called by run() before anything else, we can ensure that the destination
globals are initialized before any other dependent Javascript is
executed.

For the question of how this affects the backend code, then two things
should be mentioned:
  1. The aforementioned problem of asynchronicity-based non-determinism
  is not actually a problem for the backend. This is because on
  extension-load, gmaps_util.js was being loaded and so unless the user is
  extremely fast, s/he cannot get code that relies on gmaps_util.js'
  execution to run before gmaps_util.js fully loaded. Regardless, this
  is still an 'ugly' thing to have in the codebase.
  2. The background code called loadMapsAPI(), and therefore, to keep
  things consistent and avoid any asynchronicity issues, we moved the
  execution of the storage logic code into loadMapsAPI() by calling the new
  function from loadMapsAPI() (with an empty callback function).
The solution of using an empty callback function for #2 is kind of ugly,
but it should honestly be fine. It is the equivalent of calling a
function without caring about its callback.