Implement asynchronous listing processing
This resolves Issue #4.

This commit implements the handling of listings -- plural -- pages
asynchronously, with retry logic per listing. This accomplishes several
things:

1) Reduces the delay experienced for plural listings pages.
2) Dynamically handles arbitrarily long lists of listings pages.
  2a) Although, 'arbitrarily' should be taken with a grain of salt.

This commit only implements this functionality for streeteasy, but the
implementation for apartments.com should not be too difficult, as the
bulk of the refactoring work is done in this commit.

There has also not been much work into seeing if any code is now
dead/useless with this change. Since the multiple-listings pages used
the plural forms of the single listing page, it may be possible that the
code in the backend that was solely responsible for it is now dead. For
the content script side of things however, the plural-specific code has
remained in-use, just changed to accomodate this new approach.